requirements.txt
transformers==4.46.3
faiss-cpu==1.9.0
torch==2.5.1
sentencepiece==0.2.0
langchain==0.3.7
langchain-community==0.3.7
langchain-openai==0.2.9
beautifulsoup4==4.12.3
python-dotenv==1.0.1
langgraph==0.2.52
